requirements.txt
streamlit
langchain-community
langchain-core
langchain-text-splitters
chromadb
pypdf
ollama